Введение в SQL improved

+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -2,10 +2,12 @@ package ru.hogwarts.school.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repositories.StudentRepository;
 
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -15,8 +17,16 @@ public class StudentService {
     public StudentService(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
     }
+    public Faculty getStudentFaculty(long studentId) {
+        Student student = studentRepository.findById(studentId).orElse(null);
+        if (student != null) {
+            return student.getFaculty();
+        }
+        return null;
+    }
 
     public Student createStudent(Student student) {
+
         return studentRepository.save(student);
     }
 
@@ -25,14 +35,30 @@ public class StudentService {
         return studentRepository.findById(id).get();
     }
 
-    public Student editStudent(Student student) {
+    public Student editStudent(Student updatedStudent) {
 
-        return studentRepository.save(student);
+        Student existingStudent = studentRepository.findById(updatedStudent.getId()).orElse(null);
+
+        if (existingStudent != null) {
+
+            existingStudent.setName(updatedStudent.getName());
+            existingStudent.setAge(updatedStudent.getAge());
+
+            return studentRepository.save(existingStudent);
+        } else {
+
+            return null;
+        }
     }
 
-    public void deleteStudent(long id) {
-
-        studentRepository.deleteById(id);
+    public Student deleteStudent(long id) {
+        Optional<Student> studentOptional = studentRepository.findById(id);
+        if (studentOptional.isPresent()) {
+            studentRepository.deleteById(id);
+            return studentOptional.get();
+        } else {
+            return null;
+        }
     }
 
     public Collection<Student> getAllStudents() {
@@ -41,7 +67,6 @@ public class StudentService {
     }
 
     public Collection<Student> filterStudentsByAge(int age) {
-
         return studentRepository.findByAge(age);
     }
     public Collection<Student> findStudentsByAgeRange(int min,int max) {
